+++ src/test/resources/application.properties
@@ -1,5 +1,5 @@
 #JPA
-spring.jpa.generate-ddl=true
+spring.jpa.generate-ddl=false
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
 spring.jpa.defer-datasource-initialization=true
@@ -14,3 +14,4 @@ spring.h2.console.enabled=true
 spring.h2.console.path=/db-console
 spring.h2.console.settings.trace=false
 spring.h2.console.settings.web-allow-others=false
+spring.flyway.enabled=false
